+++ app/assets/config/manifest.js
@@ -4,7 +4,8 @@
 //= link application.js
 //= link charts.js
 //= link analytics.js
+//= link frappe-gantt.css
 //= link_tree ../images
-//= link_tree ../../javascript .js
+// = link_tree ../../javascript .js
 //= link_tree ../../../vendor/javascript .js