+++ app/assets/javascripts/charts.js
@@ -1,4 +1,5 @@
 //= require chartist
+//= require frappe-gantt
 
 function displayLessonGraph(containerId, lessonId, data) {
   new Chartist.Line(containerId, {
@@ -54,7 +55,7 @@ function displayLessonGraph(containerId, lessonId, data) {
 function displayAveragesGraph(containerId, studentId, data) {
   let responsiveOptions = [
     ['print', {
-      width: 460,
+      width: 580,
       height: 300,
       stretch: true,
       axisX:{
@@ -111,15 +112,22 @@ function displayAveragesGraph(containerId, studentId, data) {
 }
 
 function displayPercentagesGraph(containerId, data) {
+  let firstLesson = new Date(data[0].lesson_date)
+  let dayBeforeFirstLesson = new Date(firstLesson.getTime())
+  dayBeforeFirstLesson.setDate(firstLesson.getDate() - 1)
+
   let options = [
       ['', {
         height: 500
       }],
 
     ['print', {
-      width: 960,
+      width: 1200,
       height: 300,
-      stretch: true
+      stretch: true,
+      axisX: {
+        divisor: 10
+      }
     }]
   ];
 
@@ -136,16 +144,17 @@ function displayPercentagesGraph(containerId, data) {
       }
     ]
   }, {
+    fullWidth: true,
+    chartPadding: 20,
     axisX: {
       type: Chartist.FixedScaleAxis,
-      divisor: 10,
+      low: dayBeforeFirstLesson.getTime(),
+      divisor: 20,
       labelInterpolationFnc: function (value) {
         return new Date(value).toLocaleDateString('en-US'); // Convert timestamp back to date
       }
     },
     axisY: {
-      low: 0,
-      high: 100,
       labelInterpolationFnc: function (value) {
         return `${value}%`  // Convert to string with %
       }
@@ -154,4 +163,48 @@ function displayPercentagesGraph(containerId, data) {
 }
 
 
+function displayTimelineGraph(containerId, data){
+ let inMultipleYears = data.some(d => new Date(d.active_since).getFullYear() !== new Date(Date.now()).getFullYear()
+      || new Date(d.inactive_since).getFullYear()  !== new Date(Date.now()).getFullYear())
 
+ let timeline_pills = data.map( function(d, i) {
+    return {
+        id: d.student_id,
+        name: d.student_name,
+        start: new Date(d.active_since),
+        end: new Date(d.inactive_since),
+        progress: 0,
+        dependencies: d.dependent_on,
+        custom_class: 'timeline-pill'
+      }
+    }
+  )
+
+  let dataToRender = [...timeline_pills, {
+    id: 'Today',
+    name: 'Today',
+    start: new Date(),
+    end: new Date(),
+    progress: 0,
+    custom_class: 'timeline-pill today'
+  }]
+
+
+  let container = document.getElementById(containerId.substring(1))
+  // Height is set in accordance with the options for the chart below plus some extra spacing
+  // items * (padding + bar_height + one extra row for today's date) + header_height + 10
+  container.style.minHeight = `${(dataToRender.length * 25 + 50)}px`
+
+  new Gantt(containerId, dataToRender, {
+     date_format: 'YYYY-MM-DD',
+     view_mode: inMultipleYears ? 'Year' : 'Month',
+     header_height: 40,
+     padding: 12,
+     bar_height: 12,
+     bar_corner_radius: 3,
+     arrow_curve: 10,
+     popup: null,
+     gantt_start: data[0].first_lesson,
+     gantt_end: data[0].last_lesson
+  });
+}