+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added button in group view that generates report
 - Added enrolled students table for group reports
 - Show gender when viewing student tables
+- Added enrollment chart in group reports
 
 ## 0.33.0 - Lesson UUIDs & Graphs
 - Added validation for existing tags in an organization